CI: 린터를 고정하고, 스킵되던 검사를 실제로 돌게 한다
ruff가 438건으로 떨어져 있던 원인은 코드가 아니라 설정 부재 + 버전 미고정이다.
저장소에 ruff 설정이 없어 최신 ruff의 기본 룰셋이 그대로 적용되는데, uvx는 매번
최신을 받아 온다. 0.16.2가 BLE001·S110·DTZ를 기본으로 켜면서 코드를 한 줄도
안 건드린 날 CI가 깨졌다.

- pyproject.toml에 룰셋 명시. 끈 룰에는 전부 이유를 적었다. RUF001~003은
  한국어 코드베이스에서 전량 오탐이라 껐고, B904(48곳) 등은 "틀렸다"가 아니라
  "미뤘다"로 표시했다 — CI 복구 커밋에 대량 기계적 수정을 섞지 않는다
- ruff==0.16.2 핀 + CI를 uvx가 아닌 uv run ruff로 교체
- 자동 수정 209건(대부분 import 정렬), 수동 12건

린터가 잡아낸 잠복 버그 둘:
- F601: _ACTION_KEYWORD_PATTERNS에 protect_sheet 키가 중복(400·419행)
- B023: 재계획 루프의 _run_execute_once가 루프 변수 current_plan을 닫고 있었다.
  호출 위치가 밀리면 다음 회차 계획을 실행한다 — 기본 인자로 못 박았다

패키징: dev 의존성 한 줄 추가로 재빌드가 걸리자 setuptools가 scripts/·tests/까지
배포 패키지로 보고 실패했다. 캐시된 wheel 덕에 안 드러났을 뿐이다.
[tool.setuptools.packages.find]로 office_claw_sidecar*만 배포한다.

프론트엔드는 검사를 한 번도 안 받고 있었다. CI의 npm run lint --if-present가
스크립트 부재로 조용히 스킵돼, 초록불이 "통과"가 아니라 "안 함"이었다(코드의 38.9%).
- ESLint 10 flat config 추가, --if-present 제거, npm run build를 CI에 추가
- JSX 텍스트 속 맨 '>' 2건 수정(파싱 에러). esbuild도 경고하던 자리다
- ApprovalDialog의 sessionId prop→shortSession→sessionShort→hasMeta 연쇄가
  통째로 죽어 있었다. 승인 창이 세션 정보를 받아만 놓고 표시한 적이 없다

검증: ruff 통과 / pytest 878 passed·1 skipped / eslint 0 errors·45 warnings /
test:unit 40 passed. npm run build는 이 머신에서 검증 못 했다 — Node 24 +
rollup win32 네이티브가 모듈 변환 직후 죽는다(변경 전에도 동일). CI는 Node 20이다.

+++ python-sidecar/scripts/eval_ax7b_shadow.py
@@ -65,7 +65,7 @@ def _percentile(values: list[int], p: float) -> int:
     if not values:
         return 0
     sorted_values = sorted(values)
-    idx = int(round((len(sorted_values) - 1) * p))
+    idx = round((len(sorted_values) - 1) * p)
     idx = max(0, min(len(sorted_values) - 1, idx))
     return int(sorted_values[idx])
